requirements.txt
# ====================================
# SilentCodingLegend AI Agent Requirements
# ====================================

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# HTTP and API clients
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=23.0.0

# Data Models and Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
validators>=0.22.0
email-validator>=2.0.0

# AI and ML Core
llama-api-client
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
accelerate>=0.20.0
diffusers>=0.20.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Data Science and Analytics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.10.0

# Natural Language Processing
nltk>=3.8.0
tiktoken>=0.5.0

# Vector Database and Knowledge Management
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.0
networkx>=3.0.0

# Audio Processing (Optional - install only if needed)
# Uncomment the following lines if you need audio capabilities:
# System dependencies required:
# Ubuntu/Debian: sudo apt-get install portaudio19-dev python3-dev gcc
# CentOS/RHEL: sudo yum install portaudio-devel python3-devel gcc
# macOS: brew install portaudio
# SpeechRecognition>=3.10.0
# pydub>=0.25.0
# pyaudio>=0.2.11
# librosa>=0.10.0

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
moviepy>=1.0.3
imageio>=2.31.0

# Web Scraping and Parsing
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Security and Sanitization
bleach>=6.0.0

# Scheduling and Background Tasks
schedule>=1.2.0
asyncio-throttle>=1.0.0

# Development Utilities
python-dateutil>=2.8.0
typing-extensions>=4.7.0

# Development and Testing (Optional)
# Uncomment for development:
# pytest>=7.0.0
# pytest-cov>=4.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
# pre-commit>=3.0.0
# ipykernel>=6.0.0
# jupyter>=1.0.0

# Documentation (Optional)
# Uncomment for documentation building:
# sphinx>=6.0.0
# sphinx-rtd-theme>=1.0.0

# Testing Utilities (Optional)
# Uncomment for advanced testing:
# factory-boy>=3.2.0
# freezegun>=1.2.0
# responses>=0.22.0